// src/app/enums/Status.ts
export enum Status {
    ManOfStress = "Man of stress",
    TopOfTheClass = "Top of the class",
    TheLoner = "The loner",
    MommasBoy = "Momma's boy",
    TeachersPet = "Teacher's pet",
    LocalCeleb = "Local Celeb",
    Chill = "Chill",
    DumbDumb = "Dumb dumb",
    lilRebel = "lil' rebel",
    Normie = "Normie",
}